requirements.txt
websocket-client==1.7.0
pandas==2.2.0
streamlit==1.31.0
